Fix window resizing issue
Before, changing the width of the fSizeView prevented the window
from being made narrower again.

Solved by SetExplicitMinSize() of the fSizeBar (set a very small width,
so it's always possible to shrink it further) and fSizeInfo (use the max.
StringWidth as min. width) and using B_FULL_UPDATE_ON_RESIZE on the fSizeBar.
No need to set the weight of the fSizeBar anymore.

TODO:
The PathView needs to do truncating, or the window will again not be
width-decreasable if the path is long...