Split Build/BurnOutput functions
Stop using just one function to deal with the output of the
different commands (mkisofs, readcd, cdrecord) and decide what
to do with the value of fAction.

Having _BuildOutput() and _BurnOutput() also allows sending
the output to specialized parser functions. All but
ParseCdrecordLine() and ParseMkisofsLine() are non-operational.

Also renamed and unified the naming of functions over all the
Compilation*View.