Delayed image building, Update SizeBar with ISO size
Delay building when fFolderSize still missing

After choosing a folder, we get the fFolderSize of it. That may
take a while. If the user chooses to build the image without
waiting on the SizeBar to be updated, the check on the available
free space on the cache volume will always return true.

We now trigger the _Build() every second in that case and only
start building when fFolderSize is filled.

Update SizeBar with ISO size

As the resulting ISO image size can differ substantially from the
chosen folder on the harddisk (differences between the filesystem
and ISO 9660, BFS attributes, etc.), we update the SizeBar with
the size of the actual ISO image, once it has been created.