AudioView: Implement file panel to add tracks, button to play back
Use a BRefFilter to only show MIME audio/x-wav and audio/wav and
folders.
Tracks are added on top of the current selection, if no track's
selected, at the bottom.

Add a button and context menu item to play back the selected tracks.
They'll open in the preferred app for audio files. MediaPlayer
puts multiple selected tracks in a playlist.