Data, DVD: Limit disc label to 32 char (bytes)
mkisofs has a limit of 32 chars for an ID, i.e. the future disc label.

We now only allow 32 bytes in the fDiscLabel text controls and truncate
anything longer when the chosen folder name is automatically taken.

There might be an issue that 32 bytes equals not necessarily 32 chars
for certain encodings. Chinese characters, for example, can need more
than 1 byte per character.
So it might be that Chinese users have to live with shorter labels, but
I suspect that mkisofs does actually also limit to 32 bytes, not "chars"
as the error message for too long labels suggests.